chore: add tests for random user agent function
